+++ app/components/base/breadcrumbs/breadcrumbs.component.ts
@@ -1,34 +1,8 @@
-import { Component, Input, TemplateRef, ViewChild } from "@angular/core";
-import { ActivatedRoute, NavigationEnd, PRIMARY_OUTLET, Params, Router } from "@angular/router";
+import { Component, Input, OnDestroy, TemplateRef, ViewChild } from "@angular/core";
+import { Subscription } from "rxjs";
 
-interface BreadcrumbData {
-    name: string;
-    label: string;
-}
+import { Breadcrumb, BreadcrumbService } from "./breadcrumbs.service";
 
-interface Breadcrumb {
-    data: BreadcrumbData;
-    params: Params;
-    url: string;
-}
-
-interface RouteComponent {
-    name: string;
-    breadcrumb: (params: any) => BreadcrumbData;
-}
-
-function breadcrumbMethodMessage(componentName) {
-    const message = "The breadcrumb could not be generated because the route component"
-        + ` '${componentName}' doesn't have the static breadcrumb method defined`;
-    return `${message}
-    class ${componentName} {
-        // Add this method
-        public static breadcrumb(params) {
-            return {name: "Some name", label: "Some label"};
-        }
-    }
-`;
-}
 
 @Component({
     selector: "bex-crumb",
@@ -49,71 +23,16 @@ export class CrumbComponent {
     selector: "bex-breadcrumbs",
     templateUrl: "./breadcrumbs.html",
 })
-export class BreadcrumbsComponent {
+export class BreadcrumbsComponent implements OnDestroy {
     public crumbs: Breadcrumb[] = [];
 
-    constructor(private router: Router, private activatedRoute: ActivatedRoute) {
-        this.router.events.filter(event => event instanceof NavigationEnd).subscribe(event => {
-            let cls: any = this.activatedRoute.component;
-            console.log("CHnage url", event.url, cls.name);
+    private _subscription: Subscription;
 
-            // let root: ActivatedRoute = this.activatedRoute.root;
-            // this.breadcrumbs = this.getBreadcrumbs(root);
-            let root: ActivatedRoute = this.activatedRoute.root;
-            const crumb = this.getBreadcrumb(root);
-            console.log("Breadcrumb", crumb);
-            this.addBreadcrumb(crumb);
-        });
+    constructor(breadcrumbService: BreadcrumbService) {
+        this._subscription = breadcrumbService.crumbs.subscribe(x => this.crumbs = x);
     }
 
-    public addBreadcrumb(breadcrumb: Breadcrumb) {
-        this._cleanupCrumbs(breadcrumb);
-        this.crumbs.push(breadcrumb);
+    public ngOnDestroy() {
+        this._subscription.unsubscribe();
     }
-
-    private _cleanupCrumbs(breadcrumb: Breadcrumb) {
-        if (this.crumbs.length === 0) {
-            return;
-        }
-        const last = this.crumbs[this.crumbs.length - 1];
-        console.log("LASt ", last.url, "cur", breadcrumb.url);
-        if (last.url === breadcrumb.url || last.url.startsWith(breadcrumb.url)) {
-            this.crumbs.pop();
-            this._cleanupCrumbs(breadcrumb);
-        }
-    }
-
-    private getBreadcrumb(route: ActivatedRoute, url = ""): Breadcrumb {
-        if (route.children.length === 0) {
-            return null;
-        }
-
-        for (let child of route.children) {
-            if (child.outlet !== PRIMARY_OUTLET) {
-                continue;
-            }
-            let routeURL: string = child.snapshot.url.map(segment => segment.path).join("/");
-            url += `/${routeURL}`;
-            if (child.children.length === 0) {
-
-                // Found deepest child
-                const component: RouteComponent = child.snapshot.component as any;
-                let data: BreadcrumbData;
-                if (component.breadcrumb) {
-                    data = component.breadcrumb(child.snapshot.params);
-                } else {
-                    console.error(breadcrumbMethodMessage(component.name));
-                }
-                return {
-                    data,
-                    params: child.snapshot.params,
-                    url: url,
-                };
-            } else {
-                return this.getBreadcrumb(child, url);
-            }
-        }
-        return null;
-    }
-
 }
