+++ app/components/base/breadcrumbs/index.ts
@@ -1,1 +1,2 @@
 export * from "./breadcrumbs.component";
+export * from "./breadcrumbs.service";
