+++ app/components/job/details/job-details-home.component.ts
@@ -11,4 +11,7 @@ import { Component } from "@angular/core";
 })
 
 export class JobDetailsHomeComponent {
+    public static breadcrumb({id}) {
+        return { name: id, label: "Job" };
+    }
 }
